+++ package.json
@@ -3,9 +3,12 @@
   "module": "index.ts",
   "type": "module",
   "scripts": {
-    "start": "bun src/index.ts"
+    "start": "bun src/index.ts",
+    "dev": "bun --watch src/index.ts"
   },
   "devDependencies": {
+    "@types/ioredis": "^5.0.0",
+    "@types/node-cron": "^3.0.11",
     "bun-types": "latest"
   },
   "peerDependencies": {
